Correções no back-end

+++ app/src/main/java/com/example/setupbuilder/fragment/AccountFragment.kt
@@ -6,8 +6,10 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import com.example.setupbuilder.LoginActivity
 import com.example.setupbuilder.R
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
@@ -34,46 +36,43 @@ class AccountFragment : Fragment() {
             FirebaseAuth.getInstance().signOut()
             startActivity(Intent(context, LoginActivity::class.java))
         }
-                FirebaseFirestore.getInstance().collection("users").whereEqualTo("uid",user?.uid).get()
-            .addOnSuccessListener {documents ->
-                for (document in documents) {
-                    //Atualizar um campo
-//                    document.reference.update("name", "teste")
-                    home_text1?.setText("${document.data.get("name")}")
-                    name_home_input?.setText("${document.data.get("name")}")
-                }
+
+
+        FirebaseFirestore.getInstance().collection("users").document(user?.uid.toString()).get()
+            .addOnSuccessListener { document ->
+                home_text1?.setText("${document.data?.get("name")}")
+                name_home_input?.setText("${document.data?.get("name")}")
+
             }
         home_text2.setText("${user?.email}")
 
         //Editar nome
         update_button.setOnClickListener {
-            if(name_home_input.text.toString().isEmpty()){
+            if (name_home_input.text.toString().isEmpty()) {
                 name_home_input.setError("Campo vazio")
                 return@setOnClickListener
             }
 
             name_home_input.text.toString().forEach {
-                if(!it.isLetter() && !it.isWhitespace()){
+                if (!it.isLetter() && !it.isWhitespace()) {
                     name_home_input.setError("Esse campo não permite números ou caracteres especiais")
                     return@setOnClickListener
                 }
             }
 
-            FirebaseFirestore.getInstance().collection("users").whereEqualTo("uid",user?.uid).get()
-                .addOnSuccessListener {documents ->
-                    for (document in documents) {
-                        //Atualizar um campo
-                        home_text1.setText(name_home_input.text.toString())
-                        document.reference.update("name", name_home_input.text.toString())
+            FirebaseFirestore.getInstance().collection("users").document(user?.uid.toString()).get()
+                .addOnSuccessListener { document ->
+                    //Atualizar um campo
+                    home_text1.setText(name_home_input.text.toString())
+                    document.reference.update("name", name_home_input.text.toString())
 
-                    }
                 }
         }
         //Açao de deletar conta
         delete_button.setOnClickListener {
             //validar
 
-            if(password_home_input.text.toString().isEmpty()){
+            if (password_home_input.text.toString().isEmpty()) {
                 password_home_input.setError("Campo vazio.")
                 return@setOnClickListener
             }
@@ -81,39 +80,62 @@ class AccountFragment : Fragment() {
             val credential = EmailAuthProvider
                 .getCredential(user?.email.toString(), password_home_input.text.toString())
 
-            user?.reauthenticate(credential)
-                ?.addOnSuccessListener {
-                    //Se o login estiver correto
-                    user?.delete()
-                        ?.addOnCompleteListener { task ->
-                            if (task.isSuccessful) {
-                                FirebaseFirestore.getInstance().collection("users").whereEqualTo("uid",user?.uid).get()
-                                    .addOnSuccessListener {documents ->
-                                        for (document in documents) {
-                                            document.reference.delete().addOnSuccessListener {
-//                                                Toast.makeText(this, "Usuario excluido", Toast.LENGTH_LONG).show()
-//                                                startActivity(Intent(this, LoginActivity::class.java))
-                                            }
+            context?.let { it1 ->
+                MaterialAlertDialogBuilder(it1)
+                    .setTitle("Exclusão de conta")
+                    .setMessage("Tem certeza que deseja excluir sua conta? Não será possível recuperar seus dados posteriormente.")
+                    .setNeutralButton("Não") { dialog, which ->
+                        // Respond to neutral button press
+                    }
+                    .setPositiveButton("Sim") { dialog, which ->
+                        user?.reauthenticate(credential)
+                            ?.addOnSuccessListener {
+                                //Se o login estiver correto
+                                user?.delete()
+                                    ?.addOnCompleteListener { task ->
+                                        if (task.isSuccessful) {
+                                            FirebaseFirestore.getInstance().collection("users")
+                                                .document(user?.uid).get()
+                                                .addOnSuccessListener { document ->
+                                                    document.reference.delete().addOnSuccessListener {
+                                                        Toast.makeText(
+                                                            context,
+                                                            "Conta excluida",
+                                                            Toast.LENGTH_LONG
+                                                        ).show()
+                                                    }
+                                                }
+                                            FirebaseFirestore.getInstance().collection("setup")
+                                                .whereEqualTo("userUid", user?.uid).get()
+                                                .addOnSuccessListener { documents ->
+                                                    for (document in documents) {
+                                                        document.reference.delete()
+                                                    }
+                                                    startActivity(
+                                                        Intent(
+                                                            context,
+                                                            LoginActivity::class.java
+                                                        )
+                                                    )
+                                                }
+                                        } else {
+//                                Toast.makeText(this, "Erro. Reautenticação necessária", Toast.LENGTH_LONG).show()
                                         }
                                     }
-                            }else{
-//                                Toast.makeText(this, "Erro. Reautenticação necessária", Toast.LENGTH_LONG).show()
-                            }
-                        }
-                }?.addOnFailureListener {
-                    if(it.message.toString().contains("password is invalid")){
-                        password_home_input.setError("Senha incorreta.")
-                    }else {
+                            }?.addOnFailureListener {
+                                if (it.message.toString().contains("password is invalid")) {
+                                    password_home_input.setError("Senha incorreta.")
+                                } else {
 //                        Toast.makeText(this, it.message, Toast.LENGTH_LONG).show()
+                                }
+                            }
                     }
-                }
+                    .show()
+            }
+
 
         }
     }
 
-
-//        //Sair da conta
-//
-//
 
 }
